requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==2.6.0
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.4
h11==0.13.0
idna==3.3
iniconfig==1.1.1
nnfs==0.5.1
numpy==1.19.5
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyserial==3.4
PySocks==1.7.1
pytest==7.1.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.10
wsproto==1.1.0
